requirements.txt
tensorflow==2.14.0
tensorflow-recommenders==0.7.3
flask==3.0.3
mysql-connector-python==8.4.0
pandas==2.2.2
numpy==1.24.3
scikit-learn==1.5.0
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
email-validator==2.1.1
